% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootNonParametric.R
\name{bootNonParametric}
\alias{bootNonParametric}
\title{Non-Parametric Restricted Residual Resampling Bootstrap}
\usage{
bootNonParametric(
  Dependent,
  Independents,
  Repetitions,
  Hypothesis,
  rescale = FALSE
)
}
\arguments{
\item{Independents}{numeric vector or list of numeric vectors of independent variable(s)}

\item{Repetitions}{number of bootstrap repetitions to be performed}

\item{Hypothesis}{Linear restriction to be tested, must be of the format "Xi = h" where the i-th vector in the Independent
list of vectors is the variable you wish to test a restriction on and h is the restriction. Note: if you have multiple digits of independent variables,
you must enter the hypothesis with all digits, i.e. if you have 100 independent variables and you wish to test a restriction on
the 3-rd vector in the Independent list of vectors, your hypothesis should be written "X003 = h", rather than "X3 = h".}

\item{rescale}{Whether or not to rescale the residuals by (n/(n-k))^.5. This is typically done to ensure residuals have correct variance.}

\item{Dependents}{numeric vector of dependent variable}
}
\value{
A character string detailing the resulting p-value of the test.
}
\description{
This function takes a vector of a dependent variable, a vector of a single independent variable or a list of
several independent variables, a number of bootstrap repetitions, a hypothesis, and a option to rescale the residuals
and performs restricted residual resampling. To learn more about the theory, please read section 3.1 of "Bootstrap Methods
in Econometrics" by James A. MacKinnon, Queen’s Economics Department Working Paper No. 1028, 2006.
}
