% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootNonParametric.R
\name{bootNonParametric}
\alias{bootNonParametric}
\title{Non-Parametric Restricted Residual Resampling Bootstrap}
\usage{
bootNonParametric(
  dependent,
  independents,
  repetitions,
  hypothesisColumn,
  hypothesisValue,
  transform,
  replace,
  suppressConstant
)
}
\arguments{
\item{dependent}{numeric vector of dependent variable}

\item{independents}{matrix of independent variable(s), do not include constant term.}

\item{repetitions}{number of bootstrap repetitions to be performed}

\item{hypothesisColumn}{column index in 'independents' of which coefficient restriction will be tested}

\item{hypothesisValue}{hypothesized value of the coefficient of hypothesisColumn, numeric}

\item{transform}{Type of transform to apply to residuals. Acceptable values are "simple", where transform is f(u) = u(n/(n-k))^.5 (n is the number of observations and
k is the degrees of freedom),  "leverage", where transform is f(u) = u/((1-h)^.5) (h is the diagonal of the hat matrix), or "none".}

\item{replace}{Whether or not to sample residuals with replacement. Logical}

\item{suppressConstant}{logical. If true, there will be no constant term added to the regression. Defaults to false.}
}
\description{
This function takes a vector of a dependent variable, a vector of a single independent variable or a list of
several independent variables, a number of bootstrap repetitions, a hypothesis, and a option to rescale the residuals
and performs restricted residual resampling. To learn more about the theory, please read section 3.1 of "Bootstrap Methods
in Econometrics" by James A. MacKinnon, Queen’s Economics Department Working Paper No. 1028, 2006.
}
