% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootWild.R
\name{bootWild}
\alias{bootWild}
\title{Wild Bootstrap Hypothesis Testing}
\usage{
bootWild(
  dependent,
  independents,
  repetitions,
  hypothesisColumn,
  hypothesisValue,
  rv,
  transform = TRUE,
  suppressConstant = FALSE
)
}
\arguments{
\item{independents}{matrix of independent variable(s), do not include constant term.}

\item{repetitions}{number of bootstrap repetitions to be performed}

\item{hypothesisColumn}{column index in 'independents' of which coefficient restriction will be tested}

\item{hypothesisValue}{hypothesized value of the coefficient of hypothesisColumn, numeric}

\item{rv}{Distribution of random variable (with mean 0 and variance 1) to multiply the residuals by. Acceptable types are 'mammen', 'rademacher', and 'standard normal'.}

\item{transform}{Whether or not to transform the t-th residual by 1/(1-ht)^.5, where ht is the t-th diagonal of the hat matrix \code{X((X^t*X)^-1)*X^t}.
This is typically done when some observations have high leverage.}

\item{suppressConstant}{logical. If true, there will be no constant term added to the regression. Defaults to false.}

\item{dependents}{numeric vector of dependent variable}
}
\value{
A character string detailing the resulting p-value of the test.
}
\description{
This function takes a vector of a dependent variable, a vector of a single independent variable or a list of
several independent variables, a number of bootstrap repetitions, a hypothesis, and a option to rescale the residuals
and performs restricted residual resampling. To learn more about the theory, please read section 3.4 of "Bootstrap Methods
in Econometrics" by James A. MacKinnon, Queen’s Economics Department Working Paper No. 1028, 2006.
}
